[style] utils.hppにutils関係をまとめた
