+++ srcs/buffer.hpp
@@ -1,7 +1,6 @@
 #ifndef SERVER_BUFFER_HPP_
 #define SERVER_BUFFER_HPP_
 
-#include "debug.hpp" // todo: tmp
 #include <map>
 #include <string>
 #include <sys/types.h> // ssize_t
