+++ srcs/config.hpp
@@ -1,7 +1,6 @@
 #ifndef CONFIG_HPP_
 #define CONFIG_HPP_
 
-#include "debug.hpp" // todo: tmp
 #include <string>
 
 namespace config {
