+++ srcs/utils/split.cpp
@@ -1,4 +1,4 @@
-#include "split.hpp"
+#include "utils.hpp"
 
 namespace utils {
