--- srcs/utils/split.hpp
@@ -1,13 +1,0 @@
-#ifndef UTILS_SPLIT_HPP_
-#define UTILS_SPLIT_HPP_
-
-#include <string>
-#include <vector>
-
-namespace utils {
-
-std::vector<std::string> SplitStr(const std::string &src, const std::string &substring);
-
-}
-
-#endif /* UTILS_SPLIT_HPP_ */
